backend/requirements.txt
fastapi[standard]
openai 
requests 
python-dotenv
google-genai
gcloud
langchain
Pillow
firebase-admin